### 2023-11-30 #6100 @editorialbot
[PRE REVIEW]: LinguiPhyR: A Package for Linguistic Phylogenetic Analysis in R
**Submitting author:** @marccanby (<a href="http://orcid.org/0000-0002-8420-9658">Marc Canby</a>)
**Repository:** https://github.com/marccanby/linguiphyr/
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.0
