+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: linguiphyr
 Type: Package
 Title: Linguisitc Phylogenetic Analysis in R
-Version: 0.1.0
+Version: 0.2.0
 Author: Marc E. Canby
 Maintainer: Marc E. Canby <marc.canby@gmail.com>
 Description: Linguisitc Phylogenetic Analysis in R
