### 2024-01-10 #6201 @editorialbot
[REVIEW]: LinguiPhyR: A Package for Linguistic Phylogenetic Analysis in R
**Submitting author:** @marccanby (<a href="http://orcid.org/0000-0002-8420-9658">Marc Canby</a>)
**Repository:** https://github.com/marccanby/linguiphyr/
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
