### 2024-01-16 #6201 @SimonGreenhill
Some comments on the manuscript [here](https://github.com/marccanby/linguiphyr/issues/1)
