### 2024-01-16 #6201 @SimonGreenhill
The paper currently has no "Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems)". The README.md file contains this. Can this be added as a section to the paper as well?
