### 2024-01-16 #6201 @SimonGreenhill
There are no contributing guidlines -- can these be added as either a short statement to the README.md or as a CONTRIBUTIONS.md file?
