### 2024-01-16 marccanby/linguiphyr#2 @SimonGreenhill
Some issues with references
Incorrect formatting in reference list (check the bibtex {}):

- [ ] "Indo-european phylogenetics with r: "
- [ ] "indo-european"

Other issues

- [ ] Herce and Cathcart is now published, update reference details
