### 2024-01-17 35dec57
Update paper.bib - references in response to comments
