+++ paper/paper.bib
@@ -5,6 +5,23 @@
   publisher={Cambridge University Press}
 }
 
+@book{campbell2013historical,
+  title={Historical linguistics},
+  author={Campbell, Lyle},
+  year={2013},
+  publisher={Edinburgh University Press}
+}
+
+@article{heggarty2021cognacy,
+  title={Cognacy databases and phylogenetic research on Indo-European},
+  author={Heggarty, Paul},
+  journal={Annual Review of Linguistics},
+  volume={7},
+  pages={371--394},
+  year={2021},
+  publisher={Annual Reviews}
+}
+
 @book{ringeska,
   title={Historical Linguistics: Toward a Twenty-First Century Reintegration},
   author={Ringe, Don and Eska, Joseph},
