### 2024-01-17 5fd824b
Update paper.md - in response to comments
