+++ paper/paper.md
@@ -26,9 +26,9 @@ aas-journal: Astrophysical Journal <- The name of the AAS journal.--->
 
 # Introduction
 
-Phylogenetic methods have become commonplace in historical linguistics research. However, many current 
-research activities in the area are undertaken by statisticians rather than linguists, which is largely 
-(and understandably) due to the highly mathematical and computational nature of the work. This paper aims to bridge the gap 
+Phylogenetic methods have become commonplace in historical linguistics research. However, much of the work is highly 
+technical and not easily accessible to the typical classically-trained historical linguist; this is largely (and understandably) 
+due to the highly mathematical and computational nature of the work. This paper aims to bridge the gap 
 between linguistic and statistical research by introducing LinguiPhyR, an R package that provides a graphical 
 user interface (GUI) to aid in the phylogenetic analysis of linguistic data. As such, very little computational 
 or statistical expertise is required by the user. A linguist may simply upload a dataset, select 
@@ -38,35 +38,34 @@ for tree analysis are provided: a user may examine what characters are responsib
 in the tree, see the characters that are incompatible on the tree, annotate internal nodes of the tree with 
 reconstructed states, and even see a relative chronology of state changes. 
 
-We note that, at present, our software focuses on parsimony-based tree estimation and analyses.
-We make this choice because such an approach is easily interpretable by linguists: the best tree is simply
-the tree that minimizes the number of state changes. Many popular methods for phylogenetic estimation, such as maximum likelihood and 
-Bayesian inference, are less easily interpretable, as they are highly parametric and rely on a likelihood function that may be unrealistic
-or obscure to linguists. Parsimony analyses, on the other hand, make it easy to see the effect of each character in the dataset on
-tree search. Other concerns about fully parametric approaches have been raised as well, such as the suggestion that non-parametric methods like parsimony are more accurate [@barbancondiachronica2013; @tutorialNicholsWarnow; @holmes2003statistics]. 
-Nonetheless, future work will include the incorporation of other search algorithms and analytical methods into LinguiPhyR.
+We note that, at present, our software focuses on parsimony-based tree estimation and analyses. We make this choice because such an 
+approach is easily interpretable: the best tree is simply the tree that minimizes the number of state changes. 
+This makes it easy for linguists to see the effect of each character in the dataset on tree search. However, one limitation 
+of parsimony-based methods is that they are limited to searching for and analyzing tree topology: studies seeking to explore 
+ancestral node dating (glottochronology) or branch lengths are better suited to using likelihood or Bayesian approaches. 
+Future work will include the incorporation of other search algorithms and analytical methods into LinguiPhyR.
 
 # Statement of need
 
 Given the recent explosion of new linguistic phylogenetic datasets [@heggartyetal; @Tresoldi2023; @herce2023short; @jager2018global], new tools for their analyses are called for. Many linguists
 want to perform parsimony analyses of their dataset, and our software makes it easy to do so with little effort.
 In this work,
-we provide an easy-to-use tool for phylogenetic analysis that emphasizes *interpretability*, allowing linguists to understand
+we provide an easy-to-use parsimony-based tool for phylogenetic analysis that emphasizes *interpretability*, allowing linguists to understand
 why trees are returned for a particular dataset *or* what evidence a new dataset has for existing trees suggested by the community.
-<!---Currently,
-the de-facto standard for phylogenetic analysis is Bayesian inference, which, despite efforts to reduce barrier to entry,
-requires reasonable mathematical maturity to understand and operates largely as a black-box.---> 
+Currently, the go-to method for phylogenetic analysis is Bayesian inference, which, despite efforts to reduce barrier 
+to entry, requires reasonable mathematical maturity to understand and operates largely as a black-box. 
 
 The primary goals of LinguiPhyR are to
 
[…]
