### 2024-01-20 1de4e1b
Preserving case in paper.bib
