+++ paper/paper.bib
@@ -21,7 +21,7 @@
 
 @article { IndoEuropeanphylogeneticswithR,
       author = "David Goldstein",
-      title = "Indo-European phylogenetics with R: A tutorial introduction",
+      title ={{Indo-European phylogenetics with R: A tutorial introduction}},
       journal = "Indo-European Linguistics",
       year = "2020",
       publisher = "Brill",
