### 2024-01-20 ef3eb58
More case updates paper.bib
