+++ paper/paper.bib
@@ -13,7 +13,7 @@
 }
 
 @article{kelly2023traitlab,
-  title={TraitLab: a Matlab package for fitting and simulating binary tree-like data},
+  title={{TraitLab: a Matlab package for fitting and simulating binary tree-like data}},
   author={Kelly, Luke J and Nicholls, Geoff K and Ryder, Robin J and Welch, David},
   journal={arXiv preprint arXiv:2308.09060},
   year={2023}
@@ -125,7 +125,7 @@
 }
 
 @article{herce2023short,
-  title={Short vs long stem alternations in Romance verbal inflection: the S-morphome},
+  title={{Short vs long stem alternations in Romance verbal inflection: the S-morphome}},
   author={Herce Calleja, Borja and Cathcart, Chundra},
   journal={Transactions of the Philological Society},
   pages={Epub--ahead},
@@ -151,7 +151,7 @@
  journal = {Journal of Open Humanities Data},
  keyword = {en_US},
  month = {Feb},
- title = {A Global Lexical Database (GLED) for Computational Historical Linguistics},
+ title = {{A Global Lexical Database (GLED) for Computational Historical Linguistics}},
  year = {2023}
 }
 
@@ -162,7 +162,7 @@
 @article{
 heggartyetal,
 author = {Paul Heggarty  and Cormac Anderson  and Matthew Scarborough  and Benedict King  and Remco Bouckaert  and Lechosław Jocz  and Martin Joachim Kümmel  and Thomas Jügel  and Britta Irslinger  and Roland Pooth  and Henrik Liljegren  and Richard F. Strand  and Geoffrey Haig  and Martin Macák  and Ronald I. Kim  and Erik Anonby  and Tijmen Pronk  and Oleg Belyaev  and Tonya Kim Dewey-Findell  and Matthew Boutilier  and Cassandra Freiberg  and Robert Tegethoff  and Matilde Serangeli  and Nikos Liosis  and Krzysztof Stroński  and Kim Schulte  and Ganesh Kumar Gupta  and Wolfgang Haak  and Johannes Krause  and Quentin D. Atkinson  and Simon J. Greenhill  and Denise Kühnert  and Russell D. Gray },
-title = {Language trees with sampled ancestors support a hybrid model for the origin of Indo-European languages},
+title = {{Language trees with sampled ancestors support a hybrid model for the origin of Indo-European languages}},
 journal = {Science},
 volume = {381},
 number = {6656},
