+++ paper/paper.bib
@@ -124,14 +124,19 @@
 	publisher={Wiley Online Library}
 }
 
-@article{herce2023short,
-  title={{Short vs long stem alternations in Romance verbal inflection: the S-morphome}},
-  author={Herce Calleja, Borja and Cathcart, Chundra},
-  journal={Transactions of the Philological Society},
-  pages={Epub--ahead},
-  year={2023},
-  publisher={Wiley-Blackwell Publishing, Inc.}
+@article{https://doi.org/10.1111/1467-968X.12271,
+author = {Herce, Borja and Cathcart, Chundra A.},
+title = {Short vs Long Stem Alternations in Romance Verbal Inflection: The S-Morphome},
+journal = {Transactions of the Philological Society},
+volume = {n/a},
+number = {n/a},
+pages = {},
+doi = {https://doi.org/10.1111/1467-968X.12271},
+url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/1467-968X.12271},
+eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/1467-968X.12271}
 }
+
+
 
 @article{jager2018global,
   title={Global-scale phylogenetic linguistic inference from lexical resources},
