+++ paper/paper.bib
@@ -124,7 +124,7 @@
 	publisher={Wiley Online Library}
 }
 
-@article{https://doi.org/10.1111/1467-968X.12271,
+@article{herce2023short,
 author = {Herce, Borja and Cathcart, Chundra A.},
 title = {Short vs Long Stem Alternations in Romance Verbal Inflection: The S-Morphome},
 journal = {Transactions of the Philological Society},
