+++ paper/paper.bib
@@ -131,6 +131,7 @@ journal = {Transactions of the Philological Society},
 volume = {n/a},
 number = {n/a},
 pages = {},
+year={2024},
 doi = {https://doi.org/10.1111/1467-968X.12271},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/1467-968X.12271},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/1467-968X.12271}
