### 2024-01-24 marccanby/linguiphyr#2 @marccanby
I have now addressed all of these issues on the branch assigned to this issue.
