### 2024-01-24 147c663
No longer need to hit separate Upload button to upload data
