### 2024-01-25 f4f34e1
Fix UI sidebar issue, remove upload button on analysis page
