+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: linguiphyr
 Type: Package
 Title: Linguisitc Phylogenetic Analysis in R
-Version: 0.2.2
+Version: 0.2.3
 Author: Marc E. Canby
 Maintainer: Marc E. Canby <marc.canby@gmail.com>
 Description: R Shiny app for performing linguistic phylogenetic analysis in R.
