### 2024-01-25 8e00d7f
(1) fixed exhaustive mode (2) made more loading modal boxes (3) fixed height issue upon selecting a character for annotation (4) made analysis computation happen on analysis rather than search page
