+++ R/module_paup.R
@@ -17,7 +17,7 @@ paup_ui_sidebar <- function(id) {
                     checkboxInput(ns("paup_is_exhaustive"),
                                   paste0("Run in exhaustive mode ",
                                          "(requires \u2264 12 taxa)"),
-                                  TRUE))),
+                                  FALSE))),
     fluidRow(column(12,
                     radioButtons(ns("paup_keep"),
                                  NULL,
@@ -30,7 +30,7 @@ paup_ui_sidebar <- function(id) {
                      fluidRow(column(6,
                                      textInput(ns("paup_keep_score"),
                                                NULL,
-                                               428)))),
+                                               7198)))),
     conditionalPanel("!output.paup_generated",
                      fluidRow(column(6,
                                      actionButton(ns("paup_generate_nexus"),
