### 2024-01-27 6f88c30
Remove options(warn=-1) and instead suppress the benign warning in renderPlotly functions where appropriate
