### 2024-02-01 dc4aabb
Add error message when PAUP binary not found; add conditions to check that files exist before revmoving them
