### 2024-02-07 9197f41
Update error message when Paup not found; add version number to header; fix some scrolling issues.
