### 2024-02-16 ad39f08
Unit tests created along with continuous integration
