### 2024-02-16 d74de11
Remove NOTEs and WARNINGs from R CMD CHECK
