### 2024-02-16 b51911e
Remove NOTEs and WARNINGs from R CMD CHECK (#9)
