+++ R/utils_pipe.R
@@ -6,7 +6,6 @@
 #' @rdname pipe
 #' @keywords internal
 #' @importFrom magrittr %>%
-#' @usage lhs \%>\% rhs
 #' @param lhs A value or the magrittr placeholder.
 #' @param rhs A function call using the magrittr semantics.
 #' @return The result of calling `rhs(lhs)`.
@@ -20,7 +19,6 @@ NULL
 #' @name %<+%
 #' @keywords internal
 #' @importFrom ggtree %<+%
-#' @usage lhs \%<+\% rhs
 #' @param lhs A value
 #' @param rhs A value
 #' @return The result of %<+%.
