### 2024-02-16 a69b95e
Readme badge (#10)
* Remove NOTEs and WARNINGs from R CMD CHECK

* readme badge
