+++ README.md
@@ -1,7 +1,8 @@
 # LinguiphyR
 
 <!-- badges: start -->
-<!-- badges: end -->
+  [![R-CMD-check](https://github.com/marccanby/linguiphyr/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/marccanby/linguiphyr/actions/workflows/R-CMD-check.yaml)
+  <!-- badges: end -->
 
 LinguiphyR is a package that provides an R Shiny app that enables parsimony-based linguistic phylogenetic analysis. You can upload a 
 dataset of linguistic characters to the app, run a search to find optimal trees, and analyze the trees using a variety of tools
