+++ README.md
@@ -83,5 +83,5 @@ Follow these steps:
 If you have any questions or feedback regarding LinguiPhyR, or encounter a bug, please create an issue, or send an email to marc.canby@gmail.com! 
 
 We also welcome code contributions from the community. Please adhere to the following guidelines:
-* If you would like to fix a bug raised by an issue, create a pull request with your proposed changes.
+* If you notice a bug you would like to fix, first raise an issue. Then, create a pull request with your proposed changes.
 * If you are interested in suggesting a new feature, first create an issue for others to discuss before proceeding with a pull request.
