### 2024-02-19 a98558d
Edit paper to include example usage (#14)
* Add files via upload

* Update paper.md

* Delete paper/figure.png

* Delete paper/figure3.png

* Add files via upload
