### 2024-02-19 #6201 @marccanby
Responding to the unchecked boxes in Simon's review checklist:
* **Example usage:** I have added this to the paper by including a screenshot of each of the app's 3 pages (Data Upload, Tree Search, and Analysis) in the relevant sections of the paper. I also clarified in a footnote where the example dataset may be found in the repo. (All of this info is already in the README.)
* **Automated tests:** In response to another issue, I created automated tests, which run as part of continuous integration, along with R CMD Check.
* **Community guidelines:** I added a "How to contribute" section to the README, which contains contact information, and instructions on how the community may contribute.

Thanks very much for all of the feedback!
