### 2024-03-06 marccanby/linguiphyr#15 @fauxneticien
Review for JOSS
Hi @marccanby  —

Just completing my review for JOSS. Looks like Simon's review has covered quite a bit already so just some minor notes in a single issue:

1.  Permissions may not be sufficient

I ran into the "Open a Mac app from an unidentified developer" issue when I went to click "Run PAUP*" later. You might want to specify people to right click and open the `paup` executable here to trigger the `Open a Mac app from an unidentified developer` warning here after installation, so R can call it later down the line.

2. Image broken in README:
