### 2024-03-06 #6201 @fauxneticien
Two minor issues raised for @marccanby in repo: https://github.com/marccanby/linguiphyr/issues/15
