### 2024-03-27 marccanby/linguiphyr#16 @SietzeN
Review for JOSS
I have now reviewed the manuscript corresponding to the linguiphyr. I am not a linguist and the topic is a bit far from my direct expertise. Nonetheless, I provide some suggestions in the hope they will help improving the manuscript and the package.

**Installation and functionality**

- [ ] I had some difficulties with installing the package following the instructions on the readme page. The error was caused by two dependencies: ‘httpuv’ and ‘shiny’. I got the message ‘Do you want to install from sources the package which needs compilation?’. When I pressed ‘yes’ the installation failed, when I pressed ‘no’ installation was successful. You may consider adding a note about this in the readme.
- [ ] In the readme it is mentioned that the example dataset is stored in the data folder. I have now manually downloaded this file from the github page. Would it be an idea to store this dataset within the package? You could for example show in the readme how to export the built-in dataset to a local folder, or add an option in the shiny app where the user could select ‘example dataset’.

The interface itself runs smoothly and is intuitive.

**Manuscript**

- [ ] In the opening sentences of the introduction several references are made to mathematical, computational and statistical. Perhaps double-check whether all these terms are needed.
- [ ] L60. The authors write no coding is needed. While linguiphyr provides a GUI, one of the benefits of running this from within R is that users may use outputs for further analysis, or do data preparation from within R. Or are do you mean something else with ‘coding’?
- [ ] L61: ‘we believe phylogenetics can only be useful to historical linguistics if considerable analysis is given to the results of phylogenetic algorithms by linguists’
- [ ] L63. You write ‘complicated code’. Perhaps be more specific here by what you mean with this. Starting linguiphyr from within R will also require some coding, or at least some basic knowledge of programming.
- [ ] L66: “Trees are then displayed in the app and can be downloaded for inclusion in other work.” Here it might be relevant to mention interoperability with other packages. For example, how data can be prepared in other packages, or how they could be used for visualizing trees.
- [ ] L68 The author writes ‘These focus on the following questions’. The first two of these questions are also mentioned on line 53-54. I think it’s more clearly explained there. I am doubting whether a GUI approach would be most suitable for ‘making linguistic phylogenetic analysis reproducible’. If that is the aim, it might be an idea to also facilitate exporting the selected settings that were used to generate a particular tree (e.g. as a .CSV).
- [ ] L69. The sentence ‘What is the effect of particular coding’. I find this sentence not very easy to understand. By ‘coding’ do you mean how particular features are coded? Perhaps it would be good to clarify this earlier on in the manuscript (and give an example).
- [ ] L91. ‘discussion on language common in’ , perhaps rephrase to ‘terms common in’?

**Further suggestions**

- [ ] I assume the columns in the example dataset represent languages. Would it be an idea to use glottocodes or isocodes here? This would facilitate interoperability with other datasets and packages.

- [ ] References do not show up on the github page (both in-text and reference list). But this may be automatically generated when compiling the paper?

- [ ] Now the installation of PAUP goes from the browser. Would it be an idea to do that from a prompt from within R? I.e. a prompt when starting linguiphyr for the first time? It would perhaps also be helpful to add a built-in check for the latest version of PAUP and check whether this corresponds to the locally installed version.

Good luck with the revisions!
Sietze
