### 2024-03-27 #6201 @SietzeN
Hi @marccanby, I have now finished reviewing the package and manuscript. I raised some issues in repo: https://github.com/marccanby/linguiphyr/issues/16
