### 2024-04-16 067029f
Error message issue on generate nexus
