### 2024-04-16 034b328
Fix gennexus bug (#17)
* Remove NOTEs and WARNINGs from R CMD CHECK

* readme badge

* Error message issue on generate nexus

* Another small fix from same error

* Increment version number to 0.3.1
