### 2024-04-16 5cd84c5
fix regarding character being removed during polymorphism reduction in custom characters
