### 2024-05-10 f838365
Fix figure in README (#19)
