### 2024-05-10 07b67ee
Increment version number to 0.4.0
