### 2024-05-10 7c74d1d
Upload example (#20)
* Now can upload example, and download dataset

* Increment version number to 0.4.0
