### 2024-05-11 marccanby/linguiphyr#15 @marccanby
Thanks for the great advice! I have taken care of both of these points, so I am closing the issue.
