### 2024-05-17 87d68f8
Update paper.md (#24)
