### 2024-05-17 marccanby/linguiphyr#16 @marccanby
Thank you so much for all of the great advice and suggestions! I have now addressed all of these as requested. Some notes below.

Installation and functionality.
1. You are the first with this issue - I have now added a note to the README!
2. Excellent suggestion! I have now added a button in the app to use the example dataset, as well as a button to download it as a csv. So it is now all streamlined in the app and one doesn’t have to look in a folder or in the git repo.

Manuscript
1. I have reduced the usage of these terms.
2. L60: no, you’re right, this is what I mean by coding. But, I would add the point is that one doesn’t **need** to code to be able to do these analyses by using this app - prior to this, one would either have to write code to use parsimony packages in R (or similar language), or write Nexus configuration files for something like PAUP*. If one does want to perform more advanced analyses (as I have had some users ask me), we provide the option to download the Nexus configuration file, where you can then edit it by hand.
3. L61: I have rewritten this sentence to read: “While these are useful skills, giving linguists the option to spend their time analyzing trees in a GUI rather than writing code will facilitate analyses of phylogenetic inferences”
4. L63: I have removed this sentence altogether (see (3) above).
5. L66: I added “(either as images or as Nexus files)” to clarify the formats that you can download the results.
6. L68: The line about reproducibility in my draft is actually commented out, a decision I made for the exact reason you listed. So, I totally agree with your suggestion of being able to download configurations used to generate the trees so others can do the same. This is already available to some extent in the ability to download the Nexus files; however, it would be nice to do more in the future.
7. L69: I have now provided an example for “the effect of a particular coding” on l. 209, where this particular feature is explained in more detail. Thanks for the suggestion!
8. Done!

Further suggestions
1. One certainly could use glottocodes or isocodes here! It really doesn’t matter. In my work with linguists, they have either used IE clade names (e.g. “Baltic”, “Slavic” - so not actual languages per se) or often highly specific dialects of a certain language family. But if one wanted to use glottocodes or isocodes that would not be an issue.
2. Right, they show up automatically when compiling the paper, which you can see in “Actions” on the github page.
3. I have added a modal error message with the PAUP* installation instructions (more or less duplicated from the README) when PAUP* is not found - this seems to make diagnosing problems much easier. In the long run, I would definitely like to make PAUP* more streamlined within the app as you say - the PAUP* website says “Official open-source release of the program is getting very close.” - so once that’s done I hope to just tie it into the installation without requiring the extra step.
