### 2024-07-15 #6201 @ajstewartlang
:wave: @marccanby this is almost over the line - if you could address the DOI issue soon please, that would be great! Thanks.
