### 2024-07-15 #6201 @marccanby
My apologies! Will do this week!

On Mon, Jul 15, 2024 at 6:40 AM Andrew Stewart ***@***.***>
wrote:

> 👋 @marccanby <https://github.com/marccanby> this is almost over the line
> - if you could address the DOI issue soon please, that would be great!
> Thanks.
>
