### 2024-07-23 75b10b6
Update paper.bib - resolve dois (#25)
