+++ paper/paper.bib
@@ -2,14 +2,16 @@
   title={Computational phylogenetics: an introduction to designing methods for phylogeny estimation},
   author={Warnow, Tandy},
   year={2017},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
+  doi="10.1017/9781316882313"
 }
 
 @book{campbell2013historical,
-  title={Historical linguistics},
+  title={Historical Linguistics: An Introduction},
   author={Campbell, Lyle},
-  year={2013},
-  publisher={Edinburgh University Press}
+  year={2020},
+  publisher={Edinburgh University Press},
+  doi="10.1515/9781474463133"
 }
 
 @article{heggarty2021cognacy,
@@ -19,21 +21,24 @@
   volume={7},
   pages={371--394},
   year={2021},
-  publisher={Annual Reviews}
+  publisher={Annual Reviews},
+  doi="10.1146/annurev-linguistics-011619-030507"
 }
 
 @book{ringeska,
-  title={Historical Linguistics: Toward a Twenty-First Century Reintegration},
+  title={{Historical Linguistics: Toward a Twenty-First Century Reintegration}},
   author={Ringe, Don and Eska, Joseph},
   year={2013},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
+  doi="10.1017/CBO9780511980183"
 }
 
 @article{kelly2023traitlab,
   title={{TraitLab: a Matlab package for fitting and simulating binary tree-like data}},
   author={Kelly, Luke J and Nicholls, Geoff K and Ryder, Robin J and Welch, David},
   journal={arXiv preprint arXiv:2308.09060},
-  year={2023}
+  year={2023},
+  doi="10.48550/arXiv.2308.09060"
 }
 
 @article { IndoEuropeanphylogeneticswithR,
@@ -45,7 +50,7 @@
       address = "Leiden, The Netherlands",
       volume = "8",
       number = "1",
-      doi = "https://doi.org/10.1163/22125892-20201000",
+      doi = "10.1163/22125892-20201000",
       pages=      "110 - 180",
       url = "https://brill.com/view/journals/ieul/8/1/article-p110_3.xml"
 }
@@ -58,7 +63,8 @@
   number={4},
   pages={e1003537},
   year={2014},
-  publisher={Public Library of Science San Francisco, USA}
+  publisher={Public Library of Science San Francisco, USA},
+  doi="10.1371/journal.pcbi.1003537"
 }
 
 @article{maurits2017beastling,
@@ -69,7 +75,8 @@
   number={8},
   pages={e0180908},
   year={2017},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
+  doi="10.1371/journal.pone.0180908"
 }
 
 @book{swofford2002phylogenetic,
@@ -103,7 +110,8 @@
   number={2},
   pages={120--127},
   year={2003},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi="10.1111/j.1096-0031.2003.tb00299.x"
 }
 
 
@@ -115,7 +123,8 @@
   number={1},
   pages={17--32},
   year={2003},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi="10.1016/s0040-5809(02)00005-9"
 }
 
 
@@ -127,7 +136,8 @@
 	number={2},
 	pages={143--170},
 	year={2013},
-	publisher={John Benjamins}
+	publisher={John Benjamins},
+        doi="10.1075/dia.30.2.01bar"
[…]
