### 2024-07-23 a415fe9
Update paper.bib - last doi I believe (#26)
