+++ paper/paper.bib
@@ -83,7 +83,8 @@
   title={Phylogenetic analysis using parsimony ({PAUP*}) 4.0},
   author={Swofford, David L},
   year={2002},
-  publisher={Sinauer Associates: Sunderland, MA, USA}
+  publisher={Sinauer Associates: Sunderland, MA, USA},
+  doi="10.1111/j.0014-3820.2002.tb00191.x"
 }
 
 @article{tutorialNicholsWarnow,
