### 2024-09-02 b06a7db
update version for joss release
