### 2024-09-02 #6201 @marccanby
Hi there! I have now done all of this: I made a tagged release as v1.0.0; and archived the software in Zenodo: https://zenodo.org/records/13626333. The title and author list are correct. The DOI is the following: 10.5281/zenodo.13626333.

Please let me know if you need anything else to complete the publication!
