### 2024-09-02 #6201 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
