### 2024-09-19 5f4074c
Fix custom (#31)
* Update custom to take more than 10 states

* Increment version number to 1.0.1
