+++ paper/paper.md
@@ -129,7 +129,7 @@ cognate classes. Irreversible characters are binary characters that may transiti
 in the case of phonological mergers, which are generally considered irreverisble. Finally, custom characters allow the user to declare which state transitions are 
 allowed, and what the cost should be for each permitted transition. The exact way to specify this is described in the "Data Upload" page of the app.
 
-Finally, our data format supports *multi-state* and *polymorphic* characters. Multi-state characters can take more than 2 states. While it is common to convert multi-state characters into a set of independent binary traits, we recommend leaving such characters in their underlying multi-state form unless the estimation algorithm explicitly requires binary characters (for further discussion and studies on these approaches, see @rexova2003cladistic; @tutorialNicholsWarnow; @warnow2017computational). Our format also supports polymorphic characters, which are those for which a language exhibits more than one state. For lexical data, this would mean that a language manifests two cognate classes for the same semantic slot.
+Finally, our data format supports *multi-state* and *polymorphic* characters. Multi-state characters can take more than 2 states. While it is common to convert multi-state characters into a set of independent binary traits, we recommend leaving such characters in their underlying multi-state form unless the estimation algorithm explicitly requires binary characters; for further discussion on these approaches, see @rexova2003cladistic, @tutorialNicholsWarnow, @heggarty2021cognacy. Datasets may also contain polymorphic characters, which are those for which a language exhibits more than one state. For lexical data, this would mean that a language manifests two cognate classes for the same semantic slot.
 Such examples are denoted by separating the states with a */* (e.g. *1/2*) in the dataset.
 
 After uploading the data, the app presents some statistics:
