### 2024-10-03 edc47ca
Update paper.bib
