+++ paper/paper.md
@@ -129,8 +129,7 @@ cognate classes. Irreversible characters are binary characters that may transiti
 in the case of phonological mergers, which are generally considered irreverisble. Finally, custom characters allow the user to declare which state transitions are 
 allowed, and what the cost should be for each permitted transition. The exact way to specify this is described in the "Data Upload" page of the app.
 
-Finally, our data format supports *multi-state* and *polymorphic* characters. Multi-state characters can take more than 2 states. While it is common to convert multi-state characters into a set of independent binary traits, we recommend leaving such characters in their underlying multi-state form unless the estimation algorithm explicitly requires binary characters; for further discussion on these approaches, see @rexova2003cladistic, @tutorialNicholsWarnow, @heggarty2021cognacy. Datasets may also contain polymorphic characters, which are those for which a language exhibits more than one state. For lexical data, this would mean that a language manifests two cognate classes for the same semantic slot.
-Such examples are denoted by separating the states with a */* (e.g. *1/2*) in the dataset.
+Finally, our data format supports *multi-state* and *polymorphic* characters. Multi-state characters can take more than 2 states. While it is common to convert multi-state characters into a set of independent binary traits, we recommend leaving such characters in their underlying multi-state form unless the estimation algorithm explicitly requires binary characters; for further discussion on these approaches, see @rexova2003cladistic, @tutorialNicholsWarnow, @heggarty2021cognacy. Datasets may also contain polymorphic characters, which are those for which a language exhibits more than one state. For lexical data, this would mean that a language manifests two cognate classes for the same semantic slot [@canby2024]. Such examples are denoted by separating the states with a */* (e.g. *1/2*) in the dataset.
 
 After uploading the data, the app presents some statistics:
 
@@ -152,14 +151,6 @@ in the clade all share the same state, and all other languages share a different
 this condition. For example, if the dominant cognate class in a clade is lost by just one language in the clade, the character 
 will still support the grouping if the removal of the edge separating the clade from all other languages would produce a 
 less parsimonious tree. This can be examined in the "Analysis" page of the application.
-
-<!---The first five columns are special columns that provide information about each character; the remaining columns are 
-language states. The five special columns are:
-* **id:** The unique identifier for a character (typically short)
-* **feature:** The name of the character (typically a longer description)
-* **weight optional, default 1:** The (integer) weight the character should be given in phylogenetic analyses
-* **chartype optional, default 'standard':** The character type. This may be *standard*, *irreversible*, or *custom*,
-the last of which requires a special declaration format described in the app itself.--->
 
 An example usage of the "Data Upload" page is shown in \autoref{fig:figure1}; here one can see the screened dataset of @ringe2002indo uploaded.
 
@@ -194,17 +185,18 @@ sum of these support values across all edges, and minimum edge support is the mi
 
 3. **Character annotations:** The user may select any character and see the most parsimonious annotation(s) of that character's
 states across the tree (including reconstructed states at internal nodes). This is convenient for studying a character's behavior, and can help a linguist
-interpret the consequences of particular character codings on phylogeny estimation. For example, when coding the absence of a feature, a linguist has two choices: either code all languages without the feature with the same state (e.g. $0$), or code them all with different states. The former choice would suggest the absence of feature as evidence of a clade among those languages without the feature, while the latter would suggest that absence of the feature is not evidence that the languages are related. By annotating the states of each choice on proposed trees, the linguist can see the most parsimonious evolution patterns for both codings.
+interpret the consequences of particular character codings on phylogeny estimation.
[…]
