+++ paper/paper.md
@@ -94,7 +94,7 @@ the phylogenetics literature, such as *character*, *polymorphism*, and *parsimon
 
 The user first uploads a dataset of linguistic characters, which encode certain properties about languages that are likely
 to be relevant to the branching structure of the underlying tree. The characters should
-be specified in a spreadsheet and uploaded as a CSV file. An example of the data format is shown below[^1]:
+be uploaded as a CSV file. An example of the data format is shown below[^1]:
 
 | id   | feature | weight | chartype | HI | AR | GK | AL | TB | VE | AV | OC | LI | $\cdots$
 | ----------- | ----------- | ----------- | ----------- | ----------- |----------- |----------- |----------- |----------- |----------- |----------- |----------- |----------- |-----------
@@ -119,9 +119,8 @@ semantic slots (such as "bird" in the table above), and
 languages should share a state if their words for that meaning are cognate $-$ that is, the words are derived from a common ancestor via regular sound change.
 However, if a linguist can demonstrate that two languages share the same cognate due to borrowing or some other non-genetic source, then
 the languages should be given different states for that character. 
-Such cognate judgements are thus critically important to phylogenetic analysis; an abundance of phylogenetics literature discusses good methodology for doing designing characters. 
-[@ringe2002indo; @tutorialNicholsWarnow, @heggarty2021cognacy]. Classical historical linguistics references are also helpful [@ringeska, @campbell2013historical]. 
-Finally, we note that our coding scheme is applicable to phonological, morphological, and structural/typological characters, which are abundant in phylogenetic datasets.
+Such cognate judgements are thus critically important to phylogenetic analysis; an abundance of phylogenetics literature discusses good methodology for doing designing characters [@ringe2002indo, @tutorialNicholsWarnow, @heggarty2021cognacy]. Classical historical linguistics references are also helpful [@ringeska, @campbell2013historical]. 
+Finally, we note that our coding scheme is applicable to phonological, morphological, and structural/typological characters.
 
 Each character may be declared “standard”, “irreversible”, or “custom”. Standard characters permit any change of state 
 (e.g. from 0 to 1 or from 1 to 2) with uniform cost. This is generally appropriate for lexical characters where the states represent 
@@ -129,7 +128,7 @@ cognate classes. Irreversible characters are binary characters that may transiti
 in the case of phonological mergers, which are generally considered irreverisble. Finally, custom characters allow the user to declare which state transitions are 
 allowed, and what the cost should be for each permitted transition. The exact way to specify this is described in the "Data Upload" page of the app.
 
-Finally, our data format supports *multi-state* and *polymorphic* characters. Multi-state characters can take more than 2 states. While it is common to convert multi-state characters into a set of independent binary traits, we recommend leaving such characters in their underlying multi-state form unless the estimation algorithm explicitly requires binary characters; for further discussion on these approaches, see @rexova2003cladistic, @tutorialNicholsWarnow, @heggarty2021cognacy. Datasets may also contain polymorphic characters, which are those for which a language exhibits more than one state. For lexical data, this would mean that a language manifests two cognate classes for the same semantic slot [@canby2024]. Such examples are denoted by separating the states with a */* (e.g. *1/2*) in the dataset.
[…]
