+++ paper/paper.md
@@ -145,16 +145,14 @@ such a clade (a clade is a subset of languages separated from all other language
 This is computed in the strictest sense: a character only supports a hypothetical clade if the languages
 in the clade all share the same state, and all other languages share a different state.[^2]
 
-[^2]: It is important to note that a clade *on a particular tree* may be supported by more than just the characters that meet 
+[^2]: It is important to note that a clade *on a particular tree* may be supported by more than the characters that meet 
 this condition. For example, if the dominant cognate class in a clade is lost by just one language in the clade, the character 
 will still support the grouping if the removal of the edge separating the clade from all other languages would produce a 
 less parsimonious tree. This can be examined in the "Analysis" page of the application.
 
 An example usage of the "Data Upload" page is shown in \autoref{fig:figure1}; here one can see the screened dataset of @ringe2002indo uploaded.
 
-<div style="float:top; text-align:center;">
-    <img src="figure1.png" alt="Data Upload page of LinguiPhyR" style="width:80%;">
-</div>
+!["Data Upload" page of LinguiPhyR.\label{fig:figure1}](figure1.png)
 
 ## Tree Search
 
