### 2024-10-03 fe06f6a
Update paper.md
