### 2024-10-03 4f78a5a
Update paper.md
