### 2024-10-03 c9690f6
Update paper.md
