### 2024-10-03 bdc3988
Update paper.md
