+++ paper/paper.md
@@ -177,12 +177,12 @@ for further detail). To calculate total edge support and minimum edge support, w
 or support, each edge, based on whether or not the collapse of that edge would increase the parsimony score. Total edge support is the 
 sum of these support values across all edges, and minimum edge support is the minimum of these values.
 
+![Analysis page of LinguiPhyR.\label{fig:figure3}](figure3.png)
+
 2. **Character annotations:** The user may select any character and see the most parsimonious annotation(s) of that character's
 states across the tree (including inferred states at internal nodes). This is convenient for studying a character's behavior, and can help a linguist
 interpret the consequences of particular character codings on phylogeny estimation.
 <!--- For example, when coding the absence of a feature, a linguist has two choices: either code all languages without the feature with the same state (e.g. $0$), or code them all with different states. The former choice would suggest the absence of feature as evidence of a clade among those languages without the feature, while the latter would suggest that absence of the feature is not evidence that the languages are related. By annotating the states of each choice on proposed trees, the linguist can see the most parsimonious evolution patterns for both codings.--->
-
-![Analysis page of LinguiPhyR.\label{fig:figure3}](figure3.png)
 
 3. **Incompatible characters:** This reports the characters that are not compatible on a tree. This is useful
 for considering how plausible various trees are: if the set of characters that a tree is not compatible on seems unrealistic, 
