### 2024-10-04 2c7ddb4
Update paper.md
