+++ paper/paper.md
@@ -205,7 +205,7 @@ We present LinguiPhyR, a tool for analyzing phylogenetic datasets and trees via 
 programmers, LinguiPhyR can quickly enable analysis on a new linguistic dataset or provide a starting place
 for finding new trees. In our app, we especially emphasize 
 (parsimony-based) interpretability
-by providing useful visualizations and tools to see the impact of certain coding decisions on tree estimation. Future work will include the incorporation of other inference and analytical methods, such as likelihood-based or quartet approaches.
+by providing visualizations and tools to see the impact of certain coding decisions on tree estimation. Future work will incorporate other inference and analytical methods, such as likelihood-based or quartet approaches.
 
 <!---# Citations
 
