### 2024-10-04 d6e3fa0
Update paper.md
