### 2024-10-04 693fe42
Update paper.md
