### 2024-10-04 fc705ad
Update paper.md
