### 2024-10-04 ad6d426
Update paper.md (#32)
* Update paper.md

* Update paper.md

* Update paper.bib

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md
