### 2024-10-04 #6201 @marccanby
Sorry for the delay! I have reduced the length to 5 pages now, largely by fixing figure placement and making sentences more concise. Is this acceptable?
I am also happy to remove Figure 2, which I don't think shows very much, but one of the reviewers requested its inclusion, which is why it's in there.
Is there anything else I can do, or is this good to go across the finish line?
