### 2024-10-18 #6201 @samhforbes
Hi @marccanby under Canby et al there's a repeat of the https://doi.org/ prefix. I think this is the only issue I can see.
