### 2024-10-21 98e28dd
Update paper.bib (#33)
