### 2024-10-21 #6201 @marccanby
Hi @samhforbes , i just fixed the citation issue, so i think you can regenerate the pdf now. let me know if you need anything else. thanks!
