### 2024-10-22 #6201 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
